+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 > Read detailed change logs in https://github.com/sakurawald/fuji/commits/dev/
 
+- [top_chunks] fix: failed to open the GUI using `/chunks` command.
 - [top_chunks] fix: should respect the `top.rows` and `top.columns` options, while display the `Top Chunks` GUI.
 - [top_chunks] fix: should attach the `nearest players` into `chunk scores` while displaying the `Top Chunks` GUI.
 - [core] feature: simplify the language system, and enhance the `debug messages` for language system. (Make it easier to debug.)